requirements.txt
Django>=1.7
django-braces==1.4.0
